ComplectColor

+++ app/app/Services/UploadService.php
@@ -9,19 +9,20 @@ Class UploadService
 	public function store($array = [],Model $model) : Array
 	{
 		$result = [];
-		foreach ($array as $key => $itemFile)
-		{
-			if($itemFile instanceof UploadedFile)
+		if(is_array($array))
+			foreach ($array as $key => $itemFile)
 			{
+				if($itemFile instanceof UploadedFile)
+				{
 
-                if(Storage::exists($model->$key))
-                {
-                    Storage::delete($model->$key);
-                }
+	                if(Storage::exists($model->$key))
+	                {
+	                    Storage::delete($model->$key);
+	                }
 
-				$result[$key] = $itemFile->store($model->getTable().'/'.$model->id);
+					$result[$key] = $itemFile->store($model->getTable().'/'.$model->id);
+				}
 			}
-		}
 		return $result;
 	}
 }
